feat: static ws
