+++ packages/cli/src/constants.ts
@@ -3,6 +3,7 @@ export { Config } from "@dep-spy/core";
 export const defaultConfig = {
   depth: 3,
   size: false,
+  entry: null,
   output: {
     graph: "ds.graph.json",
     circularDependency: "ds.circular.json",
