+++ packages/core/src/constant.ts
@@ -1,6 +1,7 @@
 export const defaultConfig = {
   depth: 3,
   size: false,
+  entry: null,
   output: {
     graph: "ds.graph.json",
     circularDependency: "ds.circular.json",
