+++ packages/core/src/index.ts
@@ -1,4 +1,5 @@
+export { generateStaticGraph } from "./static/generateStaticGraph";
 export { generateGraph } from "./dep/generateGraph";
-export { generateGraph as generateStaticGraph } from "./dep/generateGraph";
 export type { Node, Config } from "./type";
 export type { Graph } from "./dep/graph";
+export type { StaticGraph } from "./static/staticGraph";
