+++ packages/core/src/static/generateStaticGraph.ts
@@ -1,0 +1,10 @@
+import { defaultConfig } from "../constant";
+import { Config } from "../type";
+import { StaticGraph } from "./staticGraph";
+
+export function generateStaticGraph(
+  entry: string,
+  config: Config = defaultConfig,
+): StaticGraph {
+  return new StaticGraph(entry, config);
+}
